+++ utils/test/feedback/guessing_games_test.exs
@@ -1,0 +1,5 @@
+defmodule Utils.Feedback.GuessingGamesTest do
+  use Utils.FeedbackCase,
+    module: Utils.Feedback.GuessingGames,
+    file: "../exercises/guessing_games.livemd"
+end
